skills/xitter/x-cli/pyproject.toml
[project]
name = "x-cli"
version = "0.1.0"
description = "CLI for X/Twitter API v2"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[project.scripts]
x-cli = "x_cli.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/x_cli"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = ["pytest>=8.0", "ruff>=0.4"]
